[async-http] refactor to use httpx
