+++ Makefile
@@ -7,6 +7,7 @@ default: lint typecheck test
 
 init:
 	pip install --upgrade pip
+	pip install -r requirements.txt
 	pip install -r requirements-dev.txt
 
 test:
