+++ requirements-dev.txt
@@ -8,5 +8,6 @@ flake8>=7.0.0
 python-dotenv>=1.0.1
 build>=1.2.1
 mypy>=1.9.0
-requests-mock==1.12.1
-types-requests==2.32.0.20240602
+requests-mock>=1.12.1
+respx>=0.21.1
+types-requests>=2.32.0.20240602
