+++ requirements.txt
@@ -1,0 +1,1 @@
+httpx>=0.27.0
